fix: clippy cleanup for tests/ (71 warnings across 23 files)
- needless_range_loop: convert indexed loops to iter()/iter_mut()/enumerate()
- field_reassign_with_default: restructure to struct literal with ..Default::default()
- manual RangeInclusive::contains: use (lo..=hi).contains(&x)
- doc_overindented_list_items: reduce continuation indent to 2 spaces
- clamp-like pattern: use .clamp(lo, hi)
- hex grouping: regroup hex literals to 4-digit groups
- is_multiple_of: use .is_multiple_of(n)
- map_or(false, ..): use .is_some_and(..)
- repeat().take(): use repeat_n()
- length comparison to one: use !.is_empty()
- redundant cast: remove 'as usize' on already-usize
- useless vec!: use array literal
- clone to from_ref: use std::slice::from_ref()
- identical if/else blocks: collapse to single expression
- type_complexity: factor into type alias

Skipped (pre-existing broken, do not compile):
- tests/bench_102_tilert_pipeline_goat.rs (unresolved import)
- tests/depth_invariance_micro_belief.rs (missing method audit_depth_invariance)

+++ tests/composition_279_vocab_coreset.rs
@@ -266,7 +266,7 @@ fn g2_g3_coreset_contract_and_well_formed_outputs() {
 
         assert_eq!(
             sizes.last(),
-            Some(&(N_EXPERTS as usize)),
+            Some(&N_EXPERTS),
             "G2c FAIL [{label}]: p=1.0 must select all experts, got {sizes:?}"
         );
